fixed multicast to allow different packet destination to be rewritten for different packet copies.
fixed minor issue with multicast implementation, where previously at leaf switches of a multicast tree the packet destination address (IP and MAC) were only rewritten for a single downstream switch.  this was not correct behavior when the leaf switch has more than one adjacent downstream switch.  in those scenarios, each copy of the packet has its destination IP and MAC address changed to that of its (adjacent) downstream host.

this fix was made for the tagging and counting flows and the basic versions thereof